git-svn-id: https://learning-resource-finder.googlecode.com/svn@822 59821364-4b7c-4859-2266-30ef78c0912a

+++ trunk/LearningResourceFinder/SQL/v0.3-0.4.sql
@@ -13,7 +13,7 @@ alter table resource
 
 ---- Emile 2013-09-30
 alter table resource 
-        add column counter int4;  
+        add column counter int8;  
         
 ALTER TABLE resource RENAME score  TO avgratingscore;
 ALTER TABLE resource RENAME counter  TO countrating;
